+++ src/lti_provider/urls.py
@@ -1,0 +1,8 @@
+"""URLs for the lti_provider django application."""
+from django.conf.urls import url
+
+from . import views
+
+urlpatterns = [
+    url(r"^launch$", views.launch, name="lti_provider.views.launch"),
+]